% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notebook.R
\name{add_labtools_import_to_rprofile}
\alias{add_labtools_import_to_rprofile}
\title{Add `library(labtools)` to .Rprofile}
\usage{
add_labtools_import_to_rprofile(profile_path)
}
\arguments{
\item{profile_path}{The path to the .Rprofile ro modify.}
}
\description{
Adds \code{library(labtools)} to .Rprofile to that new notebooks automatically load
\code{labtools} when opened.
}

