% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabiya_text.R
\docType{data}
\name{arabiya_text}
\alias{arabiya_text}
\title{Arabiya politics text from SANAD dataset}
\format{
A data frame with 100 rows and 1 variable:
\describe{
\item{text}{Paragraphs of Arabiya politics news}
...
}
}
\source{
\url{https://data.mendeley.com/datasets/57zpx667y9/2}
}
\usage{
arabiya_text
}
\description{
A dataset containing one hundred paragraphs of politics labelled news from
the SANAD Arabiya news data
}
\keyword{datasets}
