% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cat_sys}
\alias{cat_sys}
\title{helper function for shell output}
\usage{
cat_sys(x)
}
\arguments{
\item{x}{command to run}
}
\value{
prints output of some command
}
\description{
helper function for shell output
}
\examples{
cat_sys("ls")
}
