% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check if file exists}
\usage{
check_file_exists(path)
}
\arguments{
\item{path}{}
}
\value{
error if file doesn't exist
}
\description{
Check if file exists
}
