% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{convert_and_copy}
\alias{convert_and_copy}
\title{Copy content and their dependencies}
\usage{
convert_and_copy(from, to, copy_depend = TRUE, partial_copy = TRUE)
}
\arguments{
\item{from}{(\code{character}) The paths to content files to copy.}

\item{to}{(\code{character} of length 1) The path to where the content files and their dependencies
will be copied.}

\item{copy_depend}{(\code{logical} of length 1) If \code{FALSE}, dependencies will not be
copied.}

\item{partial_copy}{(\code{logical} of length 1) If \code{FALSE}, The entire root directory
of the content files will be copied instead of just the content files and their dependencies.}
}
\value{
(\code{character}) Paths of where the content files were copied to.
}
\description{
Copy content file files and any files they reference to a new location while preserving relative
directory locations.
Enough of the directory structure will be copied to contain all the files copies in their
original configuration.
}

