% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{copy_folder_with_links}
\alias{copy_folder_with_links}
\title{Copy folders with links}
\usage{
copy_folder_with_links(from, to)
}
\arguments{
\item{from}{(\code{character}) The path to the folder to be copied}

\item{to}{(\code{character}) Where to copy the folder to.}
}
\description{
Copies folders like \code{\link{file.copy}} except it replicates links correctly on unix-like
systems.
}

