% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{copy_note_directory}
\alias{copy_note_directory}
\title{Copy notes and their dependencies}
\usage{
copy_note_directory(path, output)
}
\arguments{
\item{path}{(\code{character}) The paths to notes to copy.}

\item{output}{(\code{character} of length 1) The path to where the notes and their dependencies
will be copied.}
}
\value{
(\code{character}) Paths of where the notes were copied to.
}
\description{
Copy note files and any files they reference to a new location.
Enough of the directory structure will be copied to contain all the files copies in their
original configuration.
}

