% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{find_readmes}
\alias{find_readmes}
\title{Find README files}
\usage{
find_readmes(path = get_project_root())
}
\arguments{
\item{path}{(\code{character} of length 1) Path somwhere under a R project in which to look for
README files.}
}
\description{
Locate all README files in the project and return their paths
}

