% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_exp}
\alias{format_exp}
\title{Formats a experiment id to a string}
\usage{
format_exp(exp)
}
\arguments{
\item{exp}{The experiment id numeric vector.}
}
\description{
Formats a experiment id numeric vector to a string for indexing.
}

