% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_rendering.R
\name{formats_quilt_can_render}
\alias{formats_quilt_can_render}
\title{File formats \code{\link{quilt}} can render}
\usage{
formats_quilt_can_render()
}
\value{
\code{character}
}
\description{
Returns a the file extensions for file formats that \code{\link{quilt}} can display on a website.
}

