% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_common_dir}
\alias{get_common_dir}
\title{Get a directory common to all paths}
\usage{
get_common_dir(paths, delim = .Platform$file.sep)
}
\arguments{
\item{paths}{(\code{character}) Input paths}

\item{delim}{(\code{character} of length 1) file path delimiter}
}
\description{
Get a directory common to all paths
}

