% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{get_config}
\alias{get_config}
\title{Get configuration options}
\usage{
get_config(option = NULL, config_path = file.path(get_project_root(),
  "config.yaml"))
}
\arguments{
\item{option}{(\code{character}) The name of the option to query}

\item{config_path}{(\code{character}) The path to the configuration file}
}
\value{
(\code{character})
}
\description{
Get configuration options
}

