% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_config_paths}
\alias{get_config_paths}
\title{Get paths to configuration files}
\usage{
get_config_paths(path, name, root, must_exist = TRUE)
}
\arguments{
\item{path}{(\code{character} of length 1) The path along which configuration files are searched.}

\item{name}{(\code{character} of length 1) The name of configuration files.}

\item{root}{(\code{character} of length 1) Where along \code{path} to start seaching. Must be a
parent directory of \code{path}.}

\item{must_exist}{(\code{logical} of length 1) If \code{FALSE}, return the paths where
configuration could be even if they dont exist.}
}
\value{
\code{character} of absolute paths.
}
\description{
Get paths to configuration files
}

