% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_config_value}
\alias{get_config_value}
\title{Get value of option in config file}
\usage{
get_config_value(path, option)
}
\arguments{
\item{path}{(\code{character} of length 1)}

\item{option}{(\code{character} of length 1)}
}
\value{
depends on content of config file. Returns \code{NA} if the option cannot be found.
}
\description{
Get the value of a specific option in a specific configuration file.
}

