% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_rendering.R
\name{get_converters}
\alias{get_converters}
\title{Get file rendering functions}
\usage{
get_converters()
}
\value{
\code{list} of \code{logical}
}
\description{
Returns a list of functions that display the content of files with HTML.
}

