% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_file_in_parent}
\alias{get_file_in_parent}
\title{Find files in parent directories}
\usage{
get_file_in_parent(path, query)
}
\arguments{
\item{path}{(\character) One or more directories in which to start looking. If multiple
directories are given then the search is started in their common parent directory.}

\item{query}{(\character) Name of the file to find.}
}
\description{
Find a file in the current or a parent directory.
}

