% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_03--configuration--get_global_options.R
\name{get_global_options}
\alias{get_global_options}
\title{Get global options values from configuration files.}
\usage{
get_global_options(main_function, sub_functions, config_path, config_name)
}
\arguments{
\item{main_function}{(\code{character} of length 1)
The name of the function to get global option names and default values from.}

\item{sub_functions}{(\code{character})
The name of the functions called by \code{main_function}.
This is used to identify invalid options in configuration files.}

\item{config_path}{(\code{character} of length 1)
The path to a configuration file or a folder that might contain one.
If a path to a folder is supplied, the \code{config_name} option is used to look for the file.}

\item{config_name}{(\code{character})
The name(s) of configuration files without file extension(s).}
}
\value{
named \code{list} of named \code{list}s
The first dimension is output types and the second a list of settings for that output type.
}
\description{
Find and parse configuration files in the given path or starting configuraiton file.
Return a two dimensional list of output types vs option values.
The `config_path` option in configuration files can be used to add additional configuration files recursivly.
}

