% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_rendering.R
\name{get_html_dependencies}
\alias{get_html_dependencies}
\title{Get html file dependencies}
\usage{
get_html_dependencies(path)
}
\arguments{
\item{path}{(\code{character}) One or more html files in which to look for references to
dependencies.}
}
\value{
\code{list} of paths are returned with elements corresponding to input \code{path}
}
\description{
Return the absolute paths of file referneced by one or more html files.
}

