% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{get_note_content_files}
\alias{get_note_content_files}
\title{Get note content file paths}
\usage{
get_note_content_files(note_path)
}
\arguments{
\item{note_path}{(\code{character} of length 1) The path to a note.}
}
\description{
Get the paths to note content that should be displayed
}

