% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{get_note_dependencies}
\alias{get_note_dependencies}
\title{Get note dependencies}
\usage{
get_note_dependencies(path, simplify = FALSE)
}
\arguments{
\item{path}{(\code{character}) One or more note files in which to look for references to
dependencies.}

\item{simplify}{(\code{logical} of length 1) If \code{FALSE}, a \code{list} of paths are returned
with elements corresponding to input directories in the \code{path} argument. If \code{TRUE}, a
single \code{character} vector is returned.}
}
\value{
Depends on the \code{simplify} option.
}
\description{
Return the absolute paths of note dependencies files.
}

