% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{get_note_files}
\alias{get_note_files}
\title{Get paths to notes}
\usage{
get_note_files(path, type = c("html"), full_names = TRUE, simplify = TRUE)
}
\arguments{
\item{path}{(\code{character}) One or more directories in which to look for note files.}

\item{type}{(\code{character}) One or more note file extensions to search for.}

\item{full_names}{(\code{logical} of length 1) \code{\link{list.files}}}

\item{simplify}{(\code{logical} of length 1) If \code{FALSE}, a \code{list} of paths are returned
with elements corresponding to input directories in the \code{path} argument. If \code{TRUE}, a
single \code{character} vector is returned.}
}
\value{
Depends on the \code{simplify} option.
}
\description{
Return the absolute paths to note files in a given directory.
}

