% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{get_note_hierarchy}
\alias{get_note_hierarchy}
\title{Get notebook note hierarchy}
\usage{
get_note_hierarchy(notebook_path = get_project_root(), implied = TRUE)
}
\arguments{
\item{notebook_path}{(\code{character} of length 1) The path to a notebook or one of its
subdirectories.}

\item{implied}{(\code{logical} of length 1) If \code{TRUE}, return parts of the hierarchy that
  no notes are assigned to, but are implied to exist (e.g. the "roots" of the hierarchy).}
}
\value{
(\code{list} of \code{character}) A list of locations in the notebook hierarchy.
}
\description{
Return a list of all the locations in the notebook note classification hierarchy.
}

