% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{get_note_paths}
\alias{get_note_paths}
\title{Gets names of notes in a notebook}
\usage{
get_note_paths(notebook_path = get_project_root(), full_names = FALSE)
}
\arguments{
\item{notebook_path}{(\code{character} of length 1) The path to a notebook or one of its
subdirectories.}

\item{full_names}{(\code{logical} of length 1) If \code{TRUE}, return the full path to each note.}
}
\description{
Returns the names/paths of notes in a given notebook.
}

