% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_option}
\alias{get_option}
\title{Get option value for a given context}
\usage{
get_option(context, option, root_context = NULL)
}
\arguments{
\item{context}{}

\item{option}{}
}
\description{
Get option value for a given context
}

