% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/organization.R
\name{get_project_root}
\alias{get_project_root}
\title{Find current project root}
\usage{
get_project_root(path = getwd())
}
\arguments{
\item{path}{(\code{character} of length 1) The path of a subdirectory of the project.}
}
\value{
The path to the current project's root
}
\description{
Finds the current project's root directory from any subdirectory by looking through the file
path for a .Proj extension.
}

