% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync.R
\name{get_remote_git_ignored}
\alias{get_remote_git_ignored}
\title{Get list of remote files ignored by git}
\usage{
get_remote_git_ignored(path, user, host, port = 22)
}
\arguments{
\item{path}{The path to a git repository on a remote file system.}

\item{user}{The user of the remote computer.}

\item{host}{The address of the remote computer.}

\item{port}{The port of ssh agent on the remote computer.}
}
\description{
Returns the list of all files ignored by git in a remote git repository.
}

