% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_rmd_yaml}
\alias{get_rmd_yaml}
\title{Extract YAML attribute}
\usage{
get_rmd_yaml(path, attribute, default = "")
}
\arguments{
\item{path}{(\code{character}) The path to a YAML file.}

\item{attribute}{(\code{character} of length 1) The key of the attribute to get.}

\item{default}{(\code{character} of length 1) the default to return of the key is not found.}
}
\description{
Gets a given attribute, based on the key, from a YAML file.
}

