% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{html_to_html}
\alias{html_to_html}
\title{Convert html to html}
\usage{
html_to_html(input, output = tempfile(fileext = ".html"))
}
\arguments{
\item{input}{(\code{character} of length 1)}

\item{output}{(\code{character} of length 1)}
}
\description{
Convert html to html
}

