% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notebook.R
\name{init_info_yaml}
\alias{init_info_yaml}
\title{Initialize notebook info}
\usage{
init_info_yaml(notebook_path)
}
\arguments{
\item{notebook_path}{The path to the notebook root directory.}
}
\description{
Initialize the info yaml file.
}

