% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipsum_text}
\alias{lipsum_text}
\title{Lipsum text from \code{stringi}}
\format{
A data frame with 100 rows and 1 variable:
\describe{
\item{text}{Paragraphs of (pseudo) random lorem ipsum text}
...
}
}
\source{
\url{https://www.rdocumentation.org/packages/stringi/versions/1.5.3/topics/stri_rand_lipsum}
}
\usage{
lipsum_text
}
\description{
A dataset containing one hundred paragraphs of lipsum text
}
\keyword{datasets}
