% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_quiltr_templates}
\alias{list_quiltr_templates}
\title{List Quiltr templates}
\usage{
list_quiltr_templates(full_path = FALSE)
}
\arguments{
\item{full_path}{(\code{logical} of length 1) If \code{TRUE}, return the full paths to templates 
instead of just their name.}
}
\value{
(\code{character}) Available Quiltr templates
}
\description{
Return available Quiltr templates. These can be made using \code{\link{make_quiltr_template}}.
}
\examples{
list_quiltr_templates()

}

