% Generated by roxygen2 (4.0.2): do not edit by hand
\name{log_time}
\alias{log_time}
\title{Keeps track of experiment start and stop times}
\usage{
log_time(exp, started = NA, finished = NA, display = TRUE)
}
\arguments{
\item{exp}{numeric vector for experiment ID. For example c(2,3) would be experiment 2-3. This should be unique.}

\item{started}{time experiment was started in YYYY-MM-DD HH:MM format.}

\item{finished}{time experiment was finished in YYYY-MM-DD HH:MM format.}

\item{display}{if TRUE, print markdown display of start and finish times.}
}
\description{
Prints and logs experiment starting and finishing times to global data.frame 'exp_time_'.
}

