% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/content_converters.R
\name{make_generic_pandoc_converter}
\alias{make_generic_pandoc_converter}
\title{Make pandoc conversion functions}
\usage{
make_generic_pandoc_converter(block_header)
}
\arguments{
\item{block_header}{(named \code{character}) The name of the programming language,
as it would appear in a fenced code block identifier. The names are the file extension for
files of that programming language.}
}
\description{
Makes generic pandoc file converter functions
}
\examples{
\dontrun{
# The following would make the function `quiltr_convert_python_to_html` and `quiltr_convert_py_to_html`
make_generic_pandoc_converter(c(py = "python", python = "python"))
}
}

