% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code-make_gh_website.R
\name{make_gh_website}
\alias{make_gh_website}
\title{Make a github pages website}
\usage{
make_gh_website(reset_branch = TRUE, commit = TRUE, clear = TRUE, ...)
}
\arguments{
\item{reset_branch}{(\code{logical} of length 1) If \code{TRUE}, change back to the current git branch
after making the webpage.}

\item{commit}{(\code{logical} of length 1) If \code{TRUE}, commit all changes made to the website
on the \code{gh-pages} branch.}

\item{clear}{(\code{logical} of length 1) If \code{TRUE}, the contents of the \code{gh-pages}
branch is deleted before the new website is copied.}

\item{...}{all other options are passed to \code{\link{make_website}}.}

\item{push}{(\code{character} of length 1) If supplied, the \code{gh-pages} branch is pushed
to the specified remote. Usually this is \code{push = "origin"}.}
}
\description{
WARNING: If you already have a website on a \code{gh-pages} branch then this function can
delete it when \code{clean = TRUE} (the defualt), so use with care.
Works like \code{\link{make_website}}, except it places the output in a \code{gh-pages} branch
of the current git repository so that it is available online at
\code{http://user_name.github.io/repository_name/}.
}

