% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{make_hierarchy_html}
\alias{make_hierarchy_html}
\title{Creates hierarchical menu html from note names}
\usage{
make_hierarchy_html(hierarchy, page_paths, notebook_name = "Notebook")
}
\arguments{
\item{notebook_path}{(\code{character} of length 1) The path to a notebook or one of its
subdirectories.}
}
\description{
Parses the names of note directories to create html for a hierarchical menu.
}

