% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{make_master_rmd}
\alias{make_master_rmd}
\title{Make composite webpage}
\usage{
make_master_rmd(name, files, location, clean = FALSE, apply_theme = TRUE)
}
\arguments{
\item{name}{(\code{character}) Name of Rmd files to use}

\item{files}{(\code{list} of \code{character}) The file in each page, corresponding to argument
\code{name}}

\item{location}{(\code{character} of length 1) Where to make the webpages}

\item{clean}{(\code{logical} of length 1) If \code{TRUE}, intermediate files are deleted after
use.}

\item{apply_theme}{(\code{logical} of length 1) If \code{TRUE}, apply parent window CSS to iframes}
}
\value{
(Named \code{character}) The file paths to created .html files named by their source Rmd
files.
}
\description{
Make webpage out of iframes.
}

