% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{make_menu_hierarchy}
\alias{make_menu_hierarchy}
\title{Creates hierarchical menu html from note names}
\usage{
make_menu_hierarchy(notes_location)
}
\arguments{
\item{notes_location}{(\code{character} of length 1) File path to the directory containing note
directories.}
}
\description{
Parses the names of note directories to create html for a hierarchical menu.
}

