% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{make_notebook_template_empty}
\alias{make_notebook_template_empty}
\title{Make empty note template directory}
\usage{
make_notebook_template_empty(location)
}
\arguments{
\item{location}{(\code{character} of length 1) Where to write the folder.}
}
\description{
Make empty note template directory
}

