% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{make_notebook_template_minimal}
\alias{make_notebook_template_minimal}
\title{Make minimal note template directory}
\usage{
make_notebook_template_minimal(location)
}
\arguments{
\item{location}{(\code{character} of length 1) Where to write the folder.}
}
\description{
Make minimal note template directory
}

