% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{make_output_yaml}
\alias{make_output_yaml}
\title{Generate _output.yaml}
\usage{
make_output_yaml()
}
\value{
(\code{character} of length 1)
}
\description{
Make the content of the file that contols knitting of the website page Rmds
}

