% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_parent_html}
\alias{make_parent_html}
\title{Make Rmd code to display HTML pages}
\usage{
make_parent_html(files, titles = NA, rmd_header = NULL, q_opt)
}
\arguments{
\item{files}{(\code{character}) One or more files to display}

\item{titles}{(\code{character} same length as \code{files}) Titles to display above embedded
content}

\item{rmd_header}{(\code{list}) YAML header data from Rmarkdown.}

\item{q_opt}{(\code{function}) The function used to get context-specific option values}
}
\value{
(\code{character} of length 1) The Rmd code to display the content provided.
}
\description{
Uses iframes to display content from other html files.
}

