% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{new_note}
\alias{new_note}
\title{Makes a new note}
\usage{
new_note(..., date = NULL, notebook = get_project_root(),
  data_folder = FALSE)
}
\arguments{
\item{...}{(\code{character}) One or more name corresponding to an organizational hierarchy.}

\item{date}{(\code{character}) Not date in the form `yyyy_mm_dd`.}

\item{notebook}{(\code{character}) The path to the notebook in which to write the note or one
of its subdirectories.}

\item{data_folder}{(\code{logical}) If \code{TRUE}, thea folder with the same name is made in
`notebook_path/data` and linked to this new note via `_data`.}
}
\description{
Make a new note for the current time and notebook.
}

