% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/content_converters.R
\name{pandoc_extension_identifiers}
\alias{pandoc_extension_identifiers}
\title{File extensions for pandoc highlighting}
\usage{
pandoc_extension_identifiers()
}
\value{
named \code{character}. Names are file extensions and values are pandoc fenced block
identifiers.
}
\description{
Returns the file extensions associated with pandoc fenced block identifiers.
}

