% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{pandoc_is_available}
\alias{pandoc_is_available}
\title{Check if pandoc exists}
\usage{
pandoc_is_available()
}
\value{
\code{logical} of length 1
}
\description{
Try to run \code{pandoc -v} and see if an error is encountered.
}

