% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_02--configuration--parse_configuration_files.R
\name{parse_configuration}
\alias{parse_configuration}
\title{Get configuration file data}
\usage{
parse_configuration(folders, option_names, config_name, global_options = NULL,
  default_path = "**")
}
\arguments{
\item{folders}{(\code{character} or named \code{list})
If \code{character}, one or more folders in which to look for configuration files to parse.
If named \code{list}, one or more R lists representing parsed configuration data.}

\item{option_names}{(\code{character} of length 1)
The set of vaild options that can be specified.
This is used to tell the differnece between options with and without path specificity}

\item{config_name}{(\code{character} of length 1)
The file name of configuration files without the file extension.}

\item{global_options}{(\code{character})
The names of global options, which should not be given path-specific values.
This is only used for error checking.
By default, no global options are assumed.}

\item{default_path}{(\code{character} of length 1)
The default path applied to options that accept path specific values but dont have a path
specified.}
}
\value{
\code{list} of \code{list} with the following items:
\describe{
  \item{option}{The name of an option for \code{function}}
  \item{value}{The value of the option}
  \item{path}{The path pattern for the option-value pair}
  \item{config_path}{The full path to the configuration file the setting was derived from}
}
}
\description{
Parse the data in configuration files for one or more folders.
}

