% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{pcr_table}
\alias{pcr_table}
\title{Creates table for PCR recipie}
\usage{
pcr_table(count, additives = c(DNA = 1), additive_concentration = rep("",
  length(additives)))
}
\arguments{
\item{count}{The number of PCR reactions.}

\item{additives}{The non-mastermix ingredients that will be added. Accepts a named numeric vector of volumes.}

\item{additive_concentration}{The concentration of non-mastermix ingredients. Accepts a named numeric vector of concentrations.}
}
\description{
Creates a markdown table for a PCR master mix recipe.
}
\examples{
pcr_table(count=8, additives=c(DNA=1, other=7))
}
\keyword{PCR}

