% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiltr.R
\name{plot.quiltr}
\alias{plot.quiltr}
\title{Plot method for quiltr object class}
\usage{
\method{plot}{quiltr}(x, sym = c("rotate", "reflect"), ...)
}
\arguments{
\item{x}{A \code{quiltr} object class.}

\item{sym}{Symmetry type. Choose from \code{rotate} or \code{reflect}.}

\item{...}{Additional graphical parameters to be passed.}
}
\description{
Plot method for quiltr object class
}
