% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiltr.R
\name{print.quiltr}
\alias{print.quiltr}
\title{Print method for quiltr object class}
\usage{
\method{print}{quiltr}(x, ...)
}
\arguments{
\item{x}{A \code{quiltr} object class.}

\item{...}{Additional parameters to be passed.}
}
\description{
Print method for quiltr object class
}
