% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{print_table}
\alias{print_table}
\title{Standardizes the printing of data.frames}
\usage{
print_table(data, column_names = NA, colors = NA, ...)
}
\arguments{
\item{data}{The data frame to be printed}

\item{column_names}{Acharacter vector of column names to be applied.}

\item{colors}{The columns of 'data' that should be colored based of their numeric content (e.g. data[, c('my_col'), drop=FALSE]).
Also accepted: A named list of vectors with names indicating columns of data and values indicating colors for the corresponding cells of 'data'.}

\item{...}{Additional key word arguments are passed to print.xtable}
}
\description{
Standardizes the printing of data.frames and allows for renaming of columns without changing original data.frame.
}

