% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{qaccess_list}
\alias{qaccess_list}
\title{See who has access to a package}
\usage{
qaccess_list(pkg)
}
\arguments{
\item{pkg}{USER/PACKAGE}
}
\description{
See who has access to a package
}
\examples{
qaccess_list("hua/iris")
}
