% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{qbuild}
\alias{qbuild}
\title{Build a package}
\usage{
qbuild(pkg, ...)
}
\arguments{
\item{pkg}{package to build}

\item{x}{data to build}
}
\value{
builds the package
}
\description{
Build a package
}
\examples{
data(iris)
qbuild("hua/iris", iris = iris)
}
