% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qinstall}
\alias{qinstall}
\title{Install a Quilt package}
\usage{
qinstall(pkg, hash = NULL, version = NULL, tag = NULL)
}
\arguments{
\item{pkg}{user/package name to install}

\item{hash}{optional}

\item{version}{optional}

\item{tag}{optional}
}
\value{
nothing, installs specified file
}
\description{
Install a Quilt package
}
\examples{
\dontrun{qinstall('akarve/examples')}
}
