% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{qload}
\alias{qload}
\title{Load file from Quilt}
\usage{
qload(pkg, file)
}
\arguments{
\item{pkg}{package name}

\item{file}{file name}
}
\value{
dataframe
}
\description{
Load file from Quilt
}
\examples{
qload("akarve/examples", "sales") \%>\% head
qload("akarve/examples", "README")
}
