% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{qlog}
\alias{qlog}
\title{Push history}
\usage{
qlog(str)
}
\arguments{
\item{str}{USER/PACKAGE}
}
\value{
push history of the package
}
\description{
Push history
}
\examples{
qlog("akarve/examples")
}
