% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qls}
\alias{qls}
\title{List installed packages}
\usage{
qls(print = FALSE)
}
\arguments{
\item{print}{print the output to console or not}
}
\value{
list of installed packages
}
\description{
List installed packages
}
\examples{
qls()
}
