% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{qparse}
\alias{qparse}
\title{Parse JSON specification to find info about a file}
\usage{
qparse(pkg, file, package_path = "~/quilt_packages/")
}
\arguments{
\item{pkg}{which package the file belongs to}

\item{file}{path to the file, see examples}

\item{package_path}{allows for user to set package location ("quilt_packages/" would set the location to the working directory)}
}
\value{
dataframe with info about a particular file
}
\description{
Parse JSON specification to find info about a file
}
\examples{
qparse("examples/wine", "quality/red")
qparse("akarve/seattle_911", "responses")
}
