% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qpeek}
\alias{qpeek}
\title{Peek at a data file}
\usage{
qpeek(str)
}
\arguments{
\item{str}{package name}
}
\value{
prints name of files
}
\description{
Roughly equivalent to typing 'examples' in Python, i.e. the package name.
}
\examples{
qpeek("akarve/examples")
}
