% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{qpush}
\alias{qpush}
\title{Push package from local to remote}
\usage{
qpush(pkg, public = FALSE, reupload = FALSE)
}
\arguments{
\item{pkg}{package to push}

\item{public}{if the package is public or not}

\item{reupload}{}
}
\value{
pushes package
}
\description{
Push package from local to remote
}
\examples{
qpush("hua/test", TRUE)
}
