% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qsearch}
\alias{qsearch}
\title{Search for packages by user or package name}
\usage{
qsearch(str)
}
\arguments{
\item{str}{user or package name}
}
\value{
list of matching packages
}
\description{
Search for packages by user or package name
}
\examples{
qsearch("akarve")
}
