% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt_form_data.R
\name{quilt_form_data}
\alias{quilt_form_data}
\title{Format text data to structure needed for \code{quilt_write_form()}}
\usage{
quilt_form_data(question, text, response_type, addID, nlow, nhigh)
}
\arguments{
\item{question}{character string: question prompt of text for labelling, e.g.: "Label this text: "}

\item{text}{character vector: text inputs for labelling}

\item{response_type}{character vector: one of c("yesno", "scale")}

\item{addID}{add an ID variable to the text data}

\item{nlow}{integer: lower end of numeric scale}

\item{nhigh}{integer: higher end of numeric scale}
}
\value{
a data.frame
}
\description{
Format text data to structure needed for \code{quilt_write_form()}
}
\examples{
\dontrun{

library(stringi)

textdat <- data.frame(text = stri_rand_lipsum(100, start_lipsum = TRUE))

qdat <- quilt_form_data(question = "Label this text: ",
text = textdat$text, response_type = "scale",
nlow = 1, nhigh = 10, addID = T)

}
}
