% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt_to_remote.R
\name{quilt_to_remote}
\alias{quilt_to_remote}
\title{Quilt and rsync result to remote server}
\usage{
quilt_to_remote(url, user, path, port = 22, just_contents = FALSE, ...)
}
\arguments{
\item{url}{(\code{character} of length 1) The url of the remote computer.}

\item{user}{(\code{character} of length 1) The user name to use on the remote computer.}

\item{path}{(\code{character} of length 1) The file path on the remote computer where the 
website will be copied.}

\item{port}{(\code{integer} of length 1) The ssh port for the remote computer.}

\item{just_contents}{\code{logical} of length 1) If \code{TRUE}, all the contents of the website
directory, instead of the directory itself, are copied to the file path specified by
\code{remote}}

\item{...}{all other options are passed to \code{\link{quilt}}.}
}
\value{
(\code{character} of length 1) The path to the local copy of the website.
}
\description{
Calls \code{\link{quilt}} and copies the output to a remote server using 
\href{https://rsync.samba.org/}{rsync}.
Ssh keys must be set up between the local and remote computers for this to work.
}

