% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt_write_form.R
\name{quilt_write_form}
\alias{quilt_write_form}
\title{Create text labelling form for importing to Qualtrics}
\usage{
quilt_write_form(
  input_data,
  page_break_every = 0,
  question_type,
  filename = "surveyOut.txt"
)
}
\arguments{
\item{input_data}{data.frame: input data generated with \code{quilt_form_data()}}

\item{page_break_every}{integer: number of text examples before page break}

\item{question_type}{character vector: one of c("dropdown", "select", "multiselect", "singleanswer", "multianswer", )}

\item{filename}{name of survey .txt form generated}
}
\value{
a .txt survey file
}
\description{
Create text labelling form for importing to Qualtrics
}
\examples{
\dontrun{

library(stringi)

textdat <- data.frame(text = stri_rand_lipsum(100, start_lipsum = TRUE))

qdat <- quilt_form_data(question = "Label this text: ", text = textdat$text,
                        response_type = "scale", nlow = 1, nhigh = 10, addID = T)

quilt_write_form(input_data = qdat,page_break_every = 1,
                 question_type = "multianswer",filename = "test/testing.txt")
}
}
