% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiltr.R
\name{quiltr}
\alias{quiltr}
\title{Pattern generator}
\usage{
quiltr(x = 3, y = 3, col = c("#AACBFF", "#E0E0E0"), ...)
}
\arguments{
\item{x}{The number of triangles for the seed pattern in the x direction.}

\item{y}{The number of triangles for the seed pattern in the y direction.}

\item{col}{A vector of colors for the pattern.}

\item{...}{Additional parameters to be passed.}
}
\value{
A list of polygon metadata
}
\description{
This function generates a pattern of isoceles right triangles
   of random colors. Patterns of initial triangle "seed" are then cloned
   and rotated based on symmetrical patterns.
}
\examples{
# Return a quiltr object
qds <- quiltr()

# Plot the object
qds <- quiltr()
plot(qds)

}
