% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_converters.R
\name{quiltr_convert_exe_to_html}
\alias{quiltr_convert_exe_to_html}
\title{Convert exe to html link}
\usage{
quiltr_convert_exe_to_html(input, output = tempfile(fileext = ".html"))
}
\arguments{
\item{input}{(\code{character} of length 1)}

\item{output}{(\code{character} of length 1)}
}
\description{
Display exe files via html link
}

