% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/content_converters.R
\name{quiltr_convert_js_to_html}
\alias{quiltr_convert_js_to_html}
\title{Convert js to html}
\usage{
quiltr_convert_js_to_html(input, output = tempfile(fileext = ".html"))
}
\arguments{
\item{input}{(\code{character} of length 1)}

\item{output}{(\code{character} of length 1)}
}
\description{
Convert java script to html
}

