% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates_default.R
\name{quiltr_template_default}
\alias{quiltr_template_default}
\title{Quiltr template for default options}
\usage{
quiltr_template_default(path = getwd(), name = "quiltr_template_default")
}
\arguments{
\item{path}{(\code{character} of length 1) Where to make the template directory.}

\item{name}{(\code{character} of length 1) The name of the template directory to make.}
}
\description{
Makes a template directory that uses default quilt options.
This is mostly for demonstration and starting new custom projects.
}

