% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{qversion}
\alias{qversion}
\title{See versions of a package}
\usage{
qversion(str)
}
\arguments{
\item{str}{USER/PACKAGE}
}
\value{
versions of the package
}
\description{
See versions of a package
}
\examples{
qversion("akarve/examples")
}
