% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_02--configuration--parse_configuration_files.R
\name{read_configuration_files}
\alias{read_configuration_files}
\title{Parse configuration files}
\usage{
read_configuration_files(folder_paths, config_name)
}
\arguments{
\item{folder_paths}{(\code{character})
The path to one or more folders from which to extract configuration file data.}

\item{config_name}{(\code{character} of length 1)
The file name of configuration files minus the file extension}
}
\value{
\code{list}
Returns \code{NA} for folders with no configuration files.
Returns \code{NULL} for empty configuration files.
}
\description{
Parse configuration files for one or more folders and return a list of their content.
}

