% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hdf5}
\alias{read_hdf5}
\title{Read HDF5 data}
\usage{
read_hdf5(h5File)
}
\arguments{
\item{h5File}{}
}
\value{
data frame
}
\description{
Helper function from \href{https://github.com/pandas-dev/pandas/issues/9636}{joschkazj}
}
