% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read_qubit}
\alias{read_qubit}
\title{Parses the output of the Qubit fluorometer}
\usage{
read_qubit(path, volume_used = NULL)
}
\arguments{
\item{path}{The file path to the output file of Qubit measurments.}

\item{volume_used}{The volume of sample added during dilution.}
}
\description{
Parses the output of the Qubit fluorometer into a dataframe. Since the Qubit stores data in the oppisite
order that the tubes are inserted, the rows order will be reversed (i.e. the first row will be the first
measured).
NOTE: if a sample is too low and has a < symbol, this function does not handle it correctly yet!
}
\keyword{Qubit}
\keyword{qubit}

