% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_02--configuration--parse_configuration_files.R
\name{reformat_configuration}
\alias{reformat_configuration}
\title{Standardize configuration content}
\usage{
reformat_configuration(raw_content, option_names)
}
\arguments{
\item{raw_content}{(\code{list})
The parsed content one or more configuration files}

\item{option_names}{(\code{character} of length 1)
The set of vaild options that can be specified.
This is used to tell the differnece between options with and without path specificity}
}
\value{
(\code{list})
}
\description{
Standardize the raw output of parsed configuration file content by applying defaults for
path-specific options. Also checks for path-spcific values being applied to global options
inappropriatly.
}

