% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/html_rendering.R
\name{render_files_to_html}
\alias{render_files_to_html}
\title{Convert a set of files to html}
\usage{
render_files_to_html(from, to, copy_depend = TRUE, partial_copy = TRUE)
}
\arguments{
\item{from}{(\code{character}) The paths to content files to copy.}

\item{to}{(\code{character} of length 1) The path to where the content files and their dependencies
will be copied.}

\item{copy_depend}{(\code{logical} of length 1) If \code{FALSE}, dependencies will not be
copied.}

\item{partial_copy}{(\code{logical} of length 1) If \code{FALSE}, The entire root directory
of the content files will be copied instead of just the content files and their dependencies.}
}
\value{
(\code{character}) Paths of where the content files were copied to.
}
\description{
Render a selection of files into HTML, preserving their relative directory structure.
Target files will be converted to HTML if they are not already HTML.
The original files and any files the HTML representations reference will be copied to a new
location.
Enough of the directory structure will be copied to allow files to have to same paths relative
to eachother as the original files.
}

