% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{rmd_to_html}
\alias{rmd_to_html}
\title{Convert rmd to html}
\usage{
rmd_to_html(input, output = tempfile(fileext = ".html"))
}
\arguments{
\item{input}{(\code{character} of length 1)}

\item{output}{(\code{character} of length 1)}
}
\description{
Convert rmd to html
}

