% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{rsync_pull}
\alias{rsync_pull}
\title{Rsync remote files ignored by git}
\usage{
rsync_pull(remote, user, host, port = 22, local = get_project_root())
}
\arguments{
\item{remote}{The path to the analogous verison of the git repository on the remote computer.}

\item{user}{The user of the remote computer.}

\item{host}{The address of the remote computer.}

\item{port}{The port of ssh agent on the remote computer.}

\item{local}{A a git repository or one of its subdirectories on the local computer.}
}
\description{
Copys files ignored by git from a remote computer using rsync.
Both computers must have rsync and git installed.
}
\details{
Will not transfer empty folders.
}

