% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/organization.R
\name{save_table}
\alias{save_table}
\title{Saves tabular data in a standard tab-separated value (.tsv) format}
\usage{
save_table(data, exp, description, ext = "tsv", path = "data")
}
\arguments{
\item{data}{A data frame to be saved}

\item{exp}{Numeric vector for experiment ID. For example c(2,3) would be experiment 2-3. This should be unique.}

\item{description}{A string describing what is to be saved. This will be converetd to lower case and spaces will be replaced by underscores.}

\item{ext}{The file extension (default: .tsv).}

\item{path}{The path to where the file will be saved (Default: data).}
}
\description{
Saves tabular data in a standard tab-separated value (.tsv) format with a standardized file name.
}

