% Generated by roxygen2 (4.0.2): do not edit by hand
\name{serial_dilution_table}
\alias{serial_dilution_table}
\title{Creates table for serial dilutions}
\usage{
serial_dilution_table(range, dilutions, volume, units = "&mu;L")
}
\arguments{
\item{range}{The inclusive range of concentrations to produce. Takes a numeric vector of length 2.}

\item{dilutions}{The number of dilutions to perform. Takes a numeric vector of length 1.}

\item{volume}{The final volume of the dilutions.}

\item{units}{The units of volume used. Default: "&mu;L".}
}
\description{
Creates a markdown table of dilution factors and final concentrations for serial dilutions given
a range of desired concentrations, the number of dilutions, and the final volume.
}
\examples{
serial_dilution_table(c(5,.00005), 5, 75)
}
\keyword{dilute}
\keyword{dilutions}
\keyword{serial}

