% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiltr.R
\name{summary.quiltr}
\alias{summary.quiltr}
\title{Summary method for \code{quiltr} object class}
\usage{
\method{summary}{quiltr}(object, ...)
}
\arguments{
\item{object}{A \code{quiltr} object class.}

\item{...}{Additional parameters to be passed.}
}
\description{
Summary method for \code{quiltr} object class
}
