% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_options.R
\name{sys_glob}
\alias{sys_glob}
\title{Wildcard expansion}
\usage{
sys_glob(path, max_search_depth = 50)
}
\arguments{
\item{path}{(\code{character} of length 1) The path to expand.}

\item{max_search_depth(\code{integer}}{of length 1) How deep to search}
}
\value{
\code{character}
}
\description{
Like \code{\link{Sys.glob}}, but also understands double wildcards for recursive matching using
double wildcards (\code{**})
}

