% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcr.R
\name{thermocycler_profile}
\alias{thermocycler_profile}
\title{Creates a thermocycler profile}
\usage{
thermocycler_profile(profile, repeats = NULL, width = NULL)
}
\arguments{
\item{profile}{(\code{list(list(vector(2)))}) A list of named 2-element numeric vectors, each
element in the list representing a stage in a profile with the 2-element numeric vector
representing the temperature and time. Optionally, a named list of the data structure described
above can be supplied to split the stages into named groups (see example).}

\item{width}{(\code{numeric}) The relative width of groupings in the graphical output. By
default, the width of each group will be proportional to the number of stage it contains.}
}
\description{
Creates a thermocycler profile in either graphical or verbal form. This is intended to be used
with arbitrary profiles. For generic PCR profiles, use \code{\link{pcr_profile}}.
}
\seealso{
\code{\link{pcr_profile}}
}

