% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notebook.R
\name{validate_readme}
\alias{validate_readme}
\title{Validate READMEs}
\usage{
validate_readme(readme_paths, missing = "warn", add = TRUE, order = TRUE)
}
\description{
Check if a readme has an entry for all files/folders in its folder.
}
\details{
NOT FINISHED
}

