% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{volume_for_dilution_table}
\alias{volume_for_dilution_table}
\title{Creates table of dilutions instructions}
\usage{
volume_for_dilution_table(initial_conc, final_volume,
  final_conc = min(initial_conc), id = names(initial_conc),
  display = TRUE, ...)
}
\arguments{
\item{initial_conc}{A numeric vector of initial concentrations of samples.}

\item{final_volume}{The desired final volume after dilution.}

\item{final_conc}{The desired final concentration. (Default: minimum of initial concentration)}

\item{id}{The row names for the samples. (Default: the names of the initial_conc vector)}

\item{display}{If True, the table is printed in markdown format.}

\item{...}{Extra arguments are passed to kable.}
}
\description{
Creates table of dilutions instructions.
}
\keyword{dilution}

