% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{copy_notes}
\alias{copy_notes}
\title{Copy notes and their dependencies}
\usage{
copy_notes(from, to, copy_depend = TRUE)
}
\arguments{
\item{from}{(\code{character}) The paths to notes to copy.}

\item{to}{(\code{character} of length 1) The path to where the notes and their dependencies
will be copied.}

\item{copy_depend}{(\code{logical} of length 1) If \code{FALSE}, dependencies will not be
copied.}
}
\value{
(\code{character}) Paths of where the notes were copied to.
}
\description{
Copy note files and any files they reference to a new location while preserving relative
directory locations.
Enough of the directory structure will be copied to contain all the files copies in their
original configuration.
}

