% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync.R
\name{get_git_ignored}
\alias{get_git_ignored}
\title{Get list of files ignored by git}
\usage{
get_git_ignored(path)
}
\arguments{
\item{path}{A a git repository or one of its subdirectories.}
}
\description{
Returns the list of all files ignored by git from anywhere within a git repository.
}

