% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gh_website.R
\name{make_gh_website}
\alias{make_gh_website}
\title{Make a github pages website}
\usage{
make_gh_website(reset_branch = TRUE, commit = TRUE, clear = TRUE,
  push = FALSE, branch = "master", remote = "origin", ...)
}
\arguments{
\item{reset_branch}{(\code{logical} of length 1) If \code{TRUE}, change back to the current git branch
after making the webpage.}

\item{commit}{(\code{logical} of length 1) If \code{TRUE}, commit all changes made to the website
on the \code{gh-pages} branch.}

\item{clear}{(\code{logical} of length 1) If \code{TRUE}, the contents of the \code{gh-pages}
branch is deleted before the new website is copied.}

\item{push}{(\code{logical} of length 1) If \code{TRUE}, push the modifications made to
the \code{gh-pages} branch to the remote specified by \code{remote}. This makes the changes 
appear online.}

\item{branch}{(\code{character} of length 1) The name of the branch in the repository that the
website source exists. This is typically 'master'.}

\item{remote}{(\code{character} of length 1) The name of the remote to push any changes to. Also
affects the inferred location of the website online.}

\item{...}{all other options are passed to \code{\link{quilt}}.}
}
\description{
WARNING: If you already have a website on a \code{gh-pages} branch then this function can 
delete it when \code{clean = TRUE} (the defualt), so use with care.
Works like \code{\link{quilt}}, except it places the output in a \code{gh-pages} branch
of the current git repository so that it is available online at
\code{http://user_name.github.io/repository_name/}.
}

