% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{make_hierarchy_html}
\alias{make_hierarchy_html}
\title{Creates hierarchical menu html from file names}
\usage{
make_hierarchy_html(hierarchy, page_paths, site_name = "Home")
}
\arguments{
\item{hierarchy}{(\code{list} of \code{character}) Locations in the menu hierarchy.}

\item{page_paths}{(\code{character}) The path to website site pages (.html files) corresponding 
argument \code{hierarchy}.}

\item{site_name}{(\code{character} of length 1) The name of the website. This is displayed 
as the link to the home page.}
}
\value{
(\code{character} of length 1) The html code to make hierarchical menu
}
\description{
Parses file paths to create html for a hierarchical menu.
}

