% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{make_parent_html}
\alias{make_parent_html}
\title{Make Rmd code to display HTML pages}
\usage{
make_parent_html(files, titles = NA, rmd_header = NULL,
  apply_theme = FALSE)
}
\arguments{
\item{files}{(\code{character}) One or more files to display}

\item{titles}{(\code{character} same length as \code{files}) Titles to display above embedded
content}

\item{rmd_header}{(\code{list}) YAML header data from Rmarkdown.}

\item{apply_theme}{(\code{logical} of length 1) If \code{TRUE}, apply parent window CSS to iframes}
}
\value{
(\code{character} of length 1) The Rmd code to display the content provided.
}
\description{
Uses iframes to display content from other html files.
}

