% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphing.R
\name{nanodrop_absorbtion_maxima}
\alias{nanodrop_absorbtion_maxima}
\title{Adds lines in a ggplot2 plot for the maximal absorption of compunds relevent to molecular biology}
\usage{
nanodrop_absorbtion_maxima()
}
\description{
Adds lines in a ggplot2 plot for the maximal absorption of compunds relevent to molecular biology.
}

