% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notebook.R
\name{new_notebook}
\alias{new_notebook}
\title{Initialize a lab notebook}
\usage{
new_notebook(location, name = "notebook", use_git = TRUE,
  use_packrat = FALSE, add_timestamp = TRUE, open = TRUE)
}
\arguments{
\item{location}{Path to new notebook location.}

\item{name}{Name of new notebook folder. Format as you would a file name.}

\item{use_git}{If \code{TRUE}, use git with the new notebook.}

\item{use_packrat}{If \code{TRUE}, use packrat with the new notebook.}

\item{add_timestamp}{If \code{TRUE}, the current date is added to the notebook name.}

\item{open}{If \code{TRUE}, the new notebook is opened in RStudio after creation (Only
 implemented in linux so far).}
}
\description{
Make the directory structure and code for a lab notebook.
}

