% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{print_table}
\alias{print_table}
\title{Standardizes the printing of data.frames}
\usage{
print_table(data, column_names = NA, colors = NA)
}
\arguments{
\item{data}{The data frame to be printed}

\item{column_names}{Acharacter vector of column names to be applied.}

\item{colors}{A named list of vectors indicating which columns should have coloring, with the elements corresponding to cells.}
}
\description{
Standardizes the printing of data.frames and allows for renaming of columns without changing original data.frame.
}

