% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{qload}
\alias{qload}
\title{Load file from Quilt}
\usage{
qload(pkg, file, file_type = "feather", package_path = "~/quilt_packages/")
}
\arguments{
\item{pkg}{package name}

\item{file}{file name}

\item{file_type}{flip between feather or csv file as the intermediary betwen parquet and R}

\item{package_path}{allows for user to set package location ("quilt_packages/" would set the location to the working directory)}
}
\value{
dataframe
}
\description{
Load file from Quilt
}
\examples{
head(qload("akarve/examples", "sales"))
qload("akarve/examples", "README")
qload("examples/wine", "quality")
qload("examples/wine", "quality/red")
qload("akarve/seattle_911", "responses")
}
