% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{qlogin}
\alias{qlogin}
\title{Login to Quilt (web interface)}
\usage{
qlogin(quilt_url = "https://pkg.quiltdata.com")
}
\arguments{
\item{quilt_url}{change if you have a separate Quilt package manager installation}
}
\value{
save auth and logs in
}
\description{
Login to Quilt (web interface)
}
\examples{
\dontrun{qlogin()}
}
