% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qpeek}
\alias{qpeek}
\title{Peek at a data file}
\usage{
qpeek(str, robust = FALSE, package_path = "~/quilt_packages/")
}
\arguments{
\item{str}{package name}

\item{robust}{more robust data, natively parsed in R}

\item{package_path}{allows for user to set package location ("quilt_packages/" would set the location to the working directory)}
}
\value{
prints name of files
}
\description{
Roughly equivalent to typing 'examples' in Python, i.e. the package name.
}
\examples{
qpeek("akarve/examples")
qpeek("akarve/examples", TRUE)
}
