% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qpeek}
\alias{qpeek}
\title{Peek at a data file}
\usage{
qpeek(pkg, robust = FALSE)
}
\arguments{
\item{pkg}{package name}

\item{robust}{more robust data, natively parsed in R}
}
\value{
prints name of files
}
\description{
Roughly equivalent to typing 'examples' in Python, i.e. the package name.
}
\examples{
qpeek("akarve/examples")
qpeek("akarve/examples", TRUE)
}
