% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recipies.R
\name{qubit_br_recipie}
\alias{qubit_br_recipie}
\title{Creates a markdown recipie for Qubit preparation}
\usage{
qubit_br_recipie(count, expected = NA, volume_added = NA,
  standards = TRUE, safety_factor = 1.1)
}
\arguments{
\item{count}{The number of samples to be measured.}

\item{expected}{The expected concentration of the DNA to be measured. Used to estimate the volume_added if
not specified.}

\item{volume_added}{The amount of sample to be diluted to 200&mu;L.}

\item{standards}{TRUE/FALSE for wether standards will be made.}

\item{safety_factor}{A factor that will be multiplied to the exact amount of working solution needed.}
}
\description{
Creates a markdown recipie for Qubit preparation.
}
\examples{
qbit_recipie(8, 5)
}
\keyword{Qubit}
\keyword{qubit}

