% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/website.R
\name{quilt}
\alias{quilt}
\title{Make a website from a directory}
\usage{
quilt(path = getwd(), output = NULL, type = formats_quilt_can_render(),
  name = "Home", clean = TRUE, overwrite = FALSE, theme = "journal",
  apply_theme = FALSE, cumulative = FALSE, use_file_names = FALSE,
  use_dir_names = TRUE, use_config_files = TRUE, name_sep = NULL,
  use_file_suffix = FALSE, use_dir_suffix = TRUE, menu_name_parser = NULL,
  note_config_name = "placement.yml",
  site_config_name = "website_build_config.yml", site_config_file = path,
  output_dir_name = "website", partial_copy = TRUE, open = TRUE)
}
\arguments{
\item{path}{(\code{character}) One or more directories in which to look for  files.}

\item{output}{(\code{character} of length 1) Location to write the output directory. The website
will be made in a directory called "website" in this location. If \code{NULL} or \code{NA}, make
the website in a temporary directory.}

\item{type}{(\code{character}) One or more file types to include in the output, specified using
file extensions without the leading dot (e.g. \code{type = c("rmd", "html")}). The order file
types are given in indicates which file is used in the case that files have the same name
but different extensions. For example, if \code{type = c("html", "rmd")} and there is a note.html
and a note.Rmd in the same directory, the note.html will be used and the note.Rmd will be
ignored.}

\item{name}{(\code{character} of length 1) The name on the link to the homepage of the website.}

\item{clean}{(\code{logical} of length 1) If \code{TRUE}, intermediate files are deleted after
use.}

\item{overwrite}{(\code{logical} of length 1) If \code{TRUE}, an existing directory with the
same name as the output directory will be overwritten.}

\item{theme}{(\code{character} of length 1) The bootstrap theme used in the website. For current
options, see \url{http://rmarkdown.rstudio.com/html_document_format.html#appearance-and-style}
\itemize{
  \item "default"
  \item "cerulean"
  \item "journal"
  \item "flatly"
  \item "readable"
  \item "spacelab"
  \item "united"
  \item "cosmo"
}}

\item{apply_theme}{(\code{logical} of length 1) If \code{TRUE}, apply notebook CSS to
content.}

\item{cumulative}{(\code{logical} of length 1) If \code{TRUE}, all of the intermendiate hierarchy
levels will be returned.}

\item{use_file_names}{(\code{logical} of length 1) If \code{TRUE}, The names of files will be
be used to determine the hierarchy.}

\item{use_dir_names}{(\code{logical} of length 1) If \code{TRUE}, The names of directories will
be used to determine the hierarchy.}

\item{use_config_files}{(\code{logical} of length 1) If \code{TRUE}, configuration
files along the notes' file path will be used to determine the hierarchy. The name of
configuration files is specified by the \code{note_config_name} option.}

\item{name_sep}{(\code{character} of length 1) A character to split file/directory names by when
using them for parts of the hierarchy.}

\item{use_file_suffix}{(\code{logical} of length 1) If \code{TRUE}, use the last part of a file
name when split by the \code{name_sep} option.}

\item{use_dir_suffix}{(\code{logical} of length 1) If \code{TRUE}, use the last part of directory
names when split by the \code{name_sep} option.}

\item{menu_name_parser}{(\code{function}) Defines a function to apply to each name in the menu hierarchy
to chenge it somehow. The function must take a single \code{character} input and output a single
\code{character}.}

\item{note_config_name}{(\code{character} of length 1) The name of content placement configuration files.}

\item{site_config_name}{(\code{character} of length 1) The name of the website building configuration file.
It does not need to exist, but if it does in a directory specified by \code{site_config_file}, it is used
automatically. To ignore website configuartion files, set this option to \code{NA} or \code{NULL}.}

\item{site_config_file}{(\code{character} of length 1) The path to a configuration file specifing
this function's option values or to the directory is is located in. The file should be in YAML format.
To ignore website configuartion files, set this option to \code{NA} or \code{NULL}.}

\item{output_dir_name}{(\code{character} of length 1) The name of the output directory.}

\item{partial_copy}{(\code{logical} of length 1) If \code{FALSE}, The entire target directory
of the notes will be copied instead of just the notes and their dependencies. It is possible that more than
just the target directory will be copied if there are files in the target directory with dependencies outside
it. Enough of the file structure will be copied to included all dependencies.}

\item{open}{(\code{logical} of length 1) If \code{TRUE}, open the newly created website in an
internet browser.}
}
\value{
(\code{character} of length 1) The file path to the created websites home page
(\code{index.html})
}
\description{
Makes a website from the contents of a directory.
Currently, only html and Rmd files are included.
All of the content will be copied and converted to html to make the website.
}

