% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_00--preface--the_quiltr_package.R
\docType{package}
\name{quiltr}
\alias{quiltr}
\alias{quiltr-package}
\title{The quiltr package}
\description{
This package includes tools for consolidating the file content of specified folders into a sharable
form (e.g. a website) in such a way that the relative file paths of the input folders/files
determine the organization the output.
In the case of website output, there are also tools to help post output on remote servers for hosting,
such as \href{https://pages.github.com/}{GitHub Pages}.
The primary function of this package is \code{\link{quilt}}; most other functions are wrappers or
components of \code{\link{quilt}}.
}
\details{
Quiltr relies heavly on other R packages, especially \code{\link{knitr}} and \code{\link{rmarkdown}}.
The source of the package is available on GitHub at \url{https://github.com/grunwaldlab/quiltr} and more
documentation is available on the GitHub Pages site at \url{http://grunwaldlab.github.io/quiltr/}.
For error reports and requests for new/changed functionality, start an
\href{https://guides.github.com/features/issues/}{issue} on the
\href{https://github.com/grunwaldlab/quiltr}{GitHub repository}.
Contributions are welcome and can be done by
\href{https://help.github.com/articles/fork-a-repo/}{forking}
the repository, making changes locally, and submitting a
\href{https://help.github.com/articles/using-pull-requests/}{pull request}.
}
\author{
Zachary Foster and Niklaus Grunwald
}

